requirements.txt
groq
playwright
python-dotenv